+++ optimus/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.7"
+__version__ = "0.12.8"
 
 
 def safe_commit() -> None:
