+++ optimus/renderer/README.md
@@ -44,9 +44,13 @@ optimus/renderer/
   visualization.py         # donut chart + hot-frames + frame-name redaction
                            # — build_donut_data, build_donut_svg,
                            #    build_hot_frames_table, redact_frame_name
+  call_tree_renderer.py    # call-tree panel (nested <details> tree)
+                           # — _render_call_tree_panel, _render_call_tree_node,
+                           #    _ct_is_user_frame, _ct_is_sql_leaf,
+                           #    _ct_is_other_frame
 ```
 
-All four submodules are imported back into `_internal.py` under their
+All five submodules are imported back into `_internal.py` under their
 original names so legacy call sites resolve unchanged. The package's
 `__init__.py` walks `dir(_internal)` and re-exports every non-dunder name
 — including underscore-prefixed internals — so external callers
@@ -145,13 +149,13 @@ is a breaking change requiring a major-version bump.
 
 ## Remaining clusters (the follow-up PR roadmap)
 
-The Plan agent's structural map flagged five more clusters in
-`_internal.py` worth extracting. Listed in recommended order
-(lowest-coupling first):
+The Plan agent's structural map flagged five clusters in `_internal.py`
+worth extracting. The first (`call_tree_renderer`) shipped in v0.12.8.
+Remaining four, listed in recommended order:
 
 | Cluster | Approx LOC | Coupling | Notes |
 |---|---|---|---|
-| `call_tree_renderer` | 240 | Weak | `_render_call_tree_panel`, `_render_call_tree_node`, `_ct_is_user_frame`, `_ct_is_sql_leaf`. Self-contained tree rendering. |
+| ✓ `call_tree_renderer` (done in v0.12.8) | 240 | Weak | `_render_call_tree_panel`, `_render_call_tree_node`, `_ct_is_user_frame`, `_ct_is_sql_leaf`, `_ct_is_other_frame`. Self-contained tree rendering. |
 | `line_drilldown` | 840 | Internal | `_render_line_drilldown_panel`, `_build_line_drilldown_callsite_index`, related helpers. Single biggest remaining chunk. Semi-public — `analyze.py` calls `_build_line_drilldown_callsite_index`. |
 | `doc_event_renderer` | 300 | Moderate | `_build_doc_event_breakdown`, `_attach_action_context`, `_extract_target_doc`. Touches the lifecycle-binding logic. |
 | `finding_enrichment` | 380 | HIGH | `_finding_to_dict`, `_attach_drilldown_chains`, `_attach_representative_callsites`, `_expand_self_time_snippets`, `_retarget_phase1_callsites_to_drilldown_leaf`, `_normalize_callsite`. Tightly coupled to `analyze.py` callers — defer until the surrounding modules are extracted and the coupling shape is clearer. |
